Update README.md
Yiannis Lymberopoulos a.k.a. YLymbE is a musician born in Athens 
Apr 1981. He is a self-learned musician, composer, and song maker.
He started active experimenting with music technology back in 2006.
Graduate from sound engineer school in 2010.And begun with debut 
Let it GO EP under the label grenade offensive in 2013.
He is a composer of many instrumental tracks and songs that he manages and 
do all the jobs of a modern independent producer.
Crazy Diamond home production as a label was an experimental work with 
Daedalus son project, PromitheasD project and Spartacus project to help him understand how digital music interacts with reality. He stopped working after the conclusion that manifestation through Art and Meditation is possible as long it follows Universal rules.
He is looking for the right fellow humans to continue researching the effects of this project and understand this way of thinking.
He enjoys life in a village near the city of Pylos Messenia and promises to continue producing music whatever the muses visit him there.